requirements.txt
python-telegram-bot==20.7
flask==3.0.3
gunicorn==22.0.0
aiosqlite==0.20.0
